# orderService/docs/api/api.yaml
openapi: 3.0.0
info:
  description: |
    This is simple client API
  version: "1.0.0"
  title: User Service
  contact:
    email: schetinnikov@gmail.com
servers:
  - url: /api/v1
tags:
  - name: user
    description: Operations about user
paths:
  /orders:
    post:
      summary: create order
      description: create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                goods:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
              required:
                - goods

      responses:
        '200':
          description: order created
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{id}:
    get:
      parameters:
        - name: id
          in: path
          description: order id
          required: true
          schema:
            type: string
      tags:
        - user

      description: Returns a order based on a serial id number
      operationId: find order by order id
      responses:
        '200':
          description: order response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:

    OrderResponse:
      type: object
      properties:
        goods:
          type: array
          items:
            $ref: '#/components/schemas/Goods'
          maxLength: 256
      required:
        - goods
      example:
        goods:
          - good_id: 10

    Goods:
      type: object
      properties:
        good_id:
          type: integer
          maxLength: 256
      required:
        - good_id
      example:
        good_id: 10

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string

    OK:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string