# orderService/internal/transport/http/order/model.gen.go
// Package order provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package order

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// Goods defines model for Goods.
type Goods struct {
	GoodId int `json:"good_id"`
}

// OrderResponse defines model for OrderResponse.
type OrderResponse struct {
	Goods []Goods `json:"goods"`
}

// CreateOrderJSONBody defines parameters for CreateOrder.
type CreateOrderJSONBody struct {
	Goods []int `json:"goods"`
}

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody CreateOrderJSONBody